fix: [MOD-1075] Plug Issue with multiple instantiations
